rssi only considers sites from ssubm
